all models running in single gpu

+++ src/models/modules/rnn_decoder.py
@@ -66,8 +66,8 @@ class LSTMAtt(nn.Module):
     def forward(self, input_feat, features, prev_word, states):
 
         if states is None:
-            states = (torch.zeros(input_feat.size(0), self.hidden_size),
-                      torch.zeros(input_feat.size(0), self.hidden_size))
+            states = (torch.zeros(input_feat.size(0), self.hidden_size).type_as(input_feat),
+                      torch.zeros(input_feat.size(0), self.hidden_size).type_as(input_feat))
 
         input = torch.cat((input_feat, prev_word), 1)
         states = self.lstmcell(input, states)
@@ -137,7 +137,7 @@ class DecoderRNN(nn.Module):
         inputs = avg_feats
         states = None
         fs = features.size(0)
-        prev_word = torch.ones(fs, 1).long() * first_token_value
+        prev_word = (torch.ones(fs, 1) * first_token_value).type_as(inputs).long()
         sampled_ids = [prev_word]
         prev_word = self.embed(prev_word).squeeze(1)
         for i in range(self.seq_length):
@@ -149,7 +149,7 @@ class DecoderRNN(nn.Module):
             if not replacement:
                 # predicted mask
                 if i == 0:
-                    predicted_mask = torch.zeros(outputs.shape).float()
+                    predicted_mask = torch.zeros(outputs.shape).type_as(outputs)
                 else:
                     batch_ind = [j for j in range(fs) if sampled_ids[i][j] != 0]
                     sampled_ids_new = sampled_ids[i][batch_ind]
